Final batch of unified_mode providers
Also converts some of them to custom resources
