+++ lib/chef/providers.rb
@@ -47,11 +47,8 @@ require_relative "provider/systemd_unit"
 require_relative "provider/template"
 require_relative "provider/user"
 require_relative "provider/whyrun_safe_ruby_block"
-require_relative "provider/windows_env"
 require_relative "provider/yum_repository"
-require_relative "provider/windows_task"
 require_relative "provider/zypper_repository"
-require_relative "provider/windows_path"
 
 require_relative "provider/package/apt"
 require_relative "provider/package/chocolatey"
