+++ lib/chef/resource/windows_package.rb
@@ -26,6 +26,7 @@ class Chef
   class Resource
     class WindowsPackage < Chef::Resource::Package
       include Chef::Mixin::Uris
+      unified_mode true
 
       provides(:windows_package) { true }
       provides :package, os: "windows"
