+++ lib/chef/resource/windows_path.rb
@@ -16,11 +16,15 @@
 # limitations under the License.
 #
 
+require_relative "../mixin/windows_env_helper" if ChefUtils.windows?
+require_relative "../mixin/wide_string"
 require_relative "../resource"
 
 class Chef
   class Resource
     class WindowsPath < Chef::Resource
+      unified_mode true
+
       provides(:windows_path) { true }
 
       description "Use the **windows_path** resource to manage the path environment variable on Microsoft Windows."
@@ -49,6 +53,40 @@ class Chef
       property :path, String,
         description: "An optional property to set the path value if it differs from the resource block's name.",
         name_property: true
+
+      action_class do
+        include Chef::Mixin::WindowsEnvHelper if ChefUtils.windows?
+      end
+
+      def load_current_resource
+        @current_resource = Chef::Resource::WindowsPath.new(new_resource.name)
+        @current_resource.path(new_resource.path)
+        @current_resource
+      end
+
+      action :add do
+        # The windows Env provider does not correctly expand variables in
+        # the PATH environment variable. Ruby expects these to be expanded.
+        #
+        path = expand_path(new_resource.path)
+        env "path" do
+          action :modify
+          delim ::File::PATH_SEPARATOR
+          value path.tr("/", '\\')
+        end
+      end
+
+      action :remove do
+        # The windows Env provider does not correctly expand variables in
+        # the PATH environment variable. Ruby expects these to be expanded.
+        #
+        path = expand_path(new_resource.path)
+        env "path" do
+          action :delete
+          delim ::File::PATH_SEPARATOR
+          value path.tr("/", '\\')
+        end
+      end
     end
   end
 end
